Bigger number for values (int to long)

+++ matrix.h
@@ -17,15 +17,15 @@ class Matrix{
   int size_x;
   int size_y;
 
-  int max_value;
+  unsigned long max_value;
 
   String result_line1;
   String result_line2;
 
   std::vector<BitArray> directions;
 
-  std::vector<unsigned short int> prev_row;
-  std::vector<unsigned short int> current_row;
+  std::vector<unsigned long> prev_row;
+  std::vector<unsigned long> current_row;
 
   std::vector< std::pair<int, int> > path_indexes;
 
@@ -38,10 +38,10 @@ class Matrix{
     bool fill();
     bool find_path();
 
-    int get(int match, int deletion, int insertion, int &direction);
-    int get_match(int x, int y);
-    int get_deletion(int x);
-    int get_insertion(int x);
+    unsigned long get(unsigned long match, unsigned long deletion, unsigned long insertion, int &direction);
+    unsigned long get_match(int x, int y);
+    unsigned long get_deletion(int x);
+    unsigned long get_insertion(int x);
 
     void print();
     void print_result();
